Create arithmetic_formatter_project.py
This is the working code for 1st project of the "scientific python" course from freeCodeCamp
the working example is on replit as well: https://replit.com/@webprice/boilerplate-arithmetic-formatter#arithmetic_arranger.py 
the code was written by twitter.com/bro_dev_
Create a function that receives a list of strings that are arithmetic problems and returns the problems arranged vertically and side-by-side.

You can access the full project description and starter code on Replit.

After going to that link, fork the project. Once you complete the project based on the instructions in 'README.md', submit your project link below.

We are still developing the interactive instructional part of the Python curriculum. For now, here are some videos on the freeCodeCamp.org YouTube channel that will teach you everything you need to know to complete this project: