Create Budget App python code raw
this is the raw version of my code for 3rd project on python's freecodecamp course:
this code was written by twitter/bro_dev_
Budget App
Create a "Category" class that can be used to create different budget categories.

You can access the full project description and starter code on Replit.

After going to that link, fork the project. Once you complete the project based on the instructions in 'README.md', submit your project link below.

We are still developing the interactive instructional part of the Python curriculum. For now, here are some videos on the freeCodeCamp.org YouTube channel that will teach you everything you need to know to complete this project: